# schemas/control-plane-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Proxy Sentinel Control Plane API
  version: 1.0.0
  description: Control-plane contract for the Proxy Sentinel frontend. The API exposes normalized events, evidence, risk snapshots, labels, shadow runs, audit logs, and session state.
servers:
  - url: /api/v1
paths:
...
  /router-observations:
    get:
      operationId: listRouterObservations
      description: Passive Huawei/H3C router observations in shadow mode; requires identity:read and never triggers enforcement.
      parameters:
        - { name: keyword, in: query, schema: { type: string } }
        - { name: ip, in: query, schema: { type: string } }
        - { name: mac, in: query, schema: { type: string } }
        - { name: vlan, in: query, schema: { type: string } }
        - { name: brand, in: query, schema: { type: string } }
        - { name: model, in: query, schema: { type: string } }
        - { name: role, in: query, description: Defaults to router; pass ap explicitly to inspect access points., schema: { type: string, enum: [router, ap, switch, firewall, endpoint, unknown], default: router } }
        - { name: status, in: query, schema: { type: string, enum: [candidate, likely, confirmed] } }
        - { name: include_candidates, in: query, description: Include low-confidence router candidates when status is not explicitly selected., schema: { type: boolean, default: false } }
        - { name: source, in: query, schema: { type: string } }
        - { name: confidence_min, in: query, schema: { type: integer, minimum: 0, maximum: 100 } }
        - { name: confidence_max, in: query, schema: { type: integer, minimum: 0, maximum: 100 } }
        - { name: first_seen_from, in: query, schema: { type: string, format: date-time } }
        - { name: first_seen_to, in: query, schema: { type: string, format: date-time } }
        - { name: last_seen_from, in: query, schema: { type: string, format: date-time } }
        - { name: last_seen_to, in: query, schema: { type: string, format: date-time } }
        - { name: infrastructure, in: query, schema: { type: boolean } }
        - { name: has_auth_binding, in: query, description: Only include observations with a current authentication session linked by exact MAC or endpoint identity., schema: { type: boolean } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: Router observation page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouterAssessmentPage' }
  /router-observations/{assessment_id}:
    get:
      operationId: getRouterObservation
      description: Passive router observation details, evidence timeline and status history; requires identity:read.
      parameters:
        - { name: assessment_id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Router observation detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouterObservationDetail' }
        '404': { description: Router observation not found }
...
  /shared-access/observations:
    get:
      operationId: listSharedBehaviorObservations
      description: Passive shared-gateway behavior observations; requires cases:read. Results are shadow-only and never authorize enforcement.
      parameters:
        - { name: view, in: query, schema: { type: string, enum: [current, history], default: current } }
        - { name: history_basis, in: query, description: 'History only. Defaults to verified multi-device evidence; clues explicitly opens weaker legacy records.', schema: { type: string, enum: [shared, clues] } }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: ip, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [candidate, likely, confirmed] } }
        - { name: coverage_state, in: query, schema: { type: string, enum: [unknown, partial, verified] } }
        - { name: confidence_min, in: query, schema: { type: integer, minimum: 0, maximum: 100 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: cursor, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        '200':
          description: Searchable shared-gateway observation page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SharedBehaviorPage' }
        '400': { description: Invalid filter or pagination }
  /shared-access/observations/{observation_id}:
    get:
      operationId: getSharedBehaviorObservation
      description: Shared-gateway score, evidence references, router correlation and immutable decision history; requires cases:read.
      parameters:
        - { name: observation_id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Shared-gateway observation detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SharedBehaviorDetail' }
        '404': { description: Shared-gateway observation not found }
...
  /application-activity/observations: {"get": {"tags": ["Applications"], "summary": "应用观测独立接口", "responses": {"200": {"description": "Success", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ApplicationObservationPage"}}}}, "503": {"description": "Application observations disabled"}}, "parameters": [{"in": "query", "name": "window", "schema": {"type": "string"}}, {"in": "query", "name": "from", "schema": {"type": "string"}}, {"in": "query", "name": "to", "schema": {"type": "string"}}, {"in": "query", "name": "sensor_id", "schema": {"type": "string"}}, {"in": "query", "name": "campus_id", "schema": {"type": "string"}}, {"in": "query", "name": "ip", "schema": {"type": "string"}}, {"in": "query", "name": "application_id", "schema": {"type": "string"}}, {"in": "query", "name": "offset", "schema": {"type": "integer", "minimum": 0}}, {"in": "query", "name": "limit", "schema": {"type": "integer", "minimum": 0}}, {"in": "query", "name": "cursor", "schema": {"type": "string"}}]}}
...
components:
  securitySchemes:
    IdentityBearer:
      type: http
      scheme: bearer
      description: Dedicated RADIUS/Portal integration credential; never a user session token.
    ProductPolicyBearer:
      type: http
      scheme: bearer
      description: Dedicated read-only product/control snapshot integration credential.
...
  schemas:
...
    SharedBehaviorScoreComponent:
      type: object
      required: [signal, score, explanation]
      properties:
        signal: { type: string, enum: [ua_os, ttl_path, tcp_stack, tls_stack, dhcp_stack, router_identity, vendor_gateway_identity] }
        score: { type: integer }
        explanation: { type: string }
    SharedBehaviorRouterContext:
      type: object
      properties:
        assessment_id: { type: string }
        brand: { type: string }
        model: { type: string }
        role: { type: string }
        status: { type: string, enum: [candidate, likely, confirmed] }
        confidence: { type: integer, minimum: 0, maximum: 100 }
        brand_attribution: { type: boolean }
    SharedBehaviorKnownDevice:
      type: object
      required: [identity_id, model, observations, first_seen, last_seen]
      properties:
        identity_id: { type: string }
        brand: { type: string }
        model: { type: string }
        os_family: { type: string }
        device_type: { type: string }
        observations: { type: integer, minimum: 2 }
        first_seen: { type: string, format: date-time }
        last_seen: { type: string, format: date-time }
...
    SharedBehaviorAssessment:
      type: object
      required: [observation_id, sensor_id, ip, status, confidence, signal_groups, reasons, conflicts, coverage_state, rule_version, first_seen, last_seen, window_start, window_end, expires_at, router, score_components, feature_samples, event_ids, known_device_count, known_devices]
      properties:
        current: { type: boolean, description: Whether this observation remains in the current complete discovery window. }
        strong_anchor: { type: string, description: Current-window multi-client evidence supporting confirmation. }
        device_lower_bound: { type: integer, minimum: 0 }
        reference_device_count_24h: { type: integer, minimum: 0, description: Historical model references; not current concurrent device count. }
        observation_id: { type: string }
        sensor_id: { type: string }
        campus_id: { type: string }
        access_domain: { type: string }
        ip: { type: string }
        endpoint_id: { type: string }
        status: { type: string, enum: [candidate, likely, confirmed] }
        confidence: { type: integer, minimum: 0, maximum: 100 }
        signal_groups: { type: array, items: { type: string } }
        reasons: { type: array, items: { type: string } }
        conflicts: { type: array, items: { type: string } }
        coverage_state: { type: string, enum: [unknown, partial, verified] }
        rule_version: { type: string }
        first_seen: { type: string, format: date-time }
        last_seen: { type: string, format: date-time }
        window_start: { type: string, format: date-time }
        window_end: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        router: { $ref: '#/components/schemas/SharedBehaviorRouterContext' }
        score_components: { type: array, items: { $ref: '#/components/schemas/SharedBehaviorScoreComponent' } }
        feature_samples:
          type: object
          additionalProperties:
            type: object
            additionalProperties: true
        event_ids: { type: array, items: { type: string } }
        known_device_count: { type: integer, minimum: 0 }
        known_device_basis: { type: string, enum: [explicit_hardware_model_lower_bound] }
        known_device_window: { type: string, enum: [24h] }
        known_devices: { type: array, items: { $ref: '#/components/schemas/SharedBehaviorKnownDevice' } }
    SharedBehaviorHistory:
      type: object
      required: [status, confidence, signal_groups, coverage_state, rule_version, observed_at, created_at]
      properties:
        status: { type: string, enum: [candidate, likely, confirmed] }
        confidence: { type: integer, minimum: 0, maximum: 100 }
        signal_groups: { type: array, items: { type: string } }
        coverage_state: { type: string, enum: [unknown, partial, verified] }
        rule_version: { type: string }
        observed_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    SharedBehaviorDetail:
      allOf:
        - { $ref: '#/components/schemas/SharedBehaviorAssessment' }
        - type: object
          required: [history]
          properties:
            history: { type: array, items: { $ref: '#/components/schemas/SharedBehaviorHistory' } }
    SharedBehaviorPage:
      type: object
      required: [items, page]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/SharedBehaviorAssessment' } }
        page:
          type: object
          required: [limit, next_cursor, total]
          properties:
            limit: { type: integer }
            next_cursor: { type: [string, 'null'] }
            total: { type: integer }
...
    ApplicationMatch: {"type": "object", "properties": {"rule_id": {"type": "string"}, "target_id": {"type": "string"}, "target_type": {"type": "string"}, "name": {"type": "string"}, "category": {"type": "string"}, "source": {"type": "string"}, "source_version": {"type": "string"}, "confidence": {"type": "number"}}}
    ApplicationObservation: {"type": "object", "properties": {"event_id": {"type": "string"}, "timestamp": {"type": "string"}, "sensor_id": {"type": "string"}, "campus_id": {"type": "string"}, "ip": {"type": "string"}, "connection_id": {"type": "string"}, "event_type": {"type": "string"}, "domain": {"type": "string"}, "source_field": {"type": "string"}, "bundle_version": {"type": "string"}, "match": {"$ref": "#/components/schemas/ApplicationMatch"}, "upload_bytes": {"type": "integer", "minimum": 0, "nullable": true}, "download_bytes": {"type": "integer", "minimum": 0, "nullable": true}}}
...
    ApplicationObservationPage: {"type": "object", "properties": {"items": {"type": "array", "items": {"$ref": "#/components/schemas/ApplicationObservation"}}, "total": {"type": "integer", "minimum": 0}, "total_as_of": {"type": "string", "format": "date-time", "description": "分页总数的统计来源时间；数据库模式最多延迟五秒。"}, "limit": {"type": "integer", "minimum": 0}, "offset": {"type": "integer", "minimum": 0}, "next_cursor": {"type": "string", "description": "下一页游标；为空表示没有下一页。游标不能与非零 offset 同用。"}}}
...
    RouterEvidence:
      type: object
      required: [evidence_id, assessment_id, kind, source, source_family, strength, score, rule_id, rule_version, explanation, association_quality, first_seen, last_seen, expires_at, expired, event_ids]
      properties:
        evidence_id: { type: string }
        assessment_id: { type: string }
        kind: { type: string, enum: [router_signal, conflict, confirmed_router] }
        endpoint_id: { type: string }
        ip: { type: string }
        mac: { type: string }
        vlan: { type: string }
        brand: { type: string }
        series: { type: string }
        model: { type: string }
        role: { type: string }
        source: { type: string }
        source_family: { type: string }
        source_event_type: { type: string }
        sensor_id: { type: string }
        shared_observation_id: { type: string }
        raw_value: { type: string }
        strength: { type: string, enum: [weak, medium, strong] }
        score: { type: integer }
        rule_id: { type: string }
        rule_version: { type: string }
        explanation: { type: string }
        conflict: { type: boolean }
        conflict_code: { type: string }
        exclusion: { type: boolean }
        brand_reference_only: { type: boolean }
        association_quality: { type: string }
        association_reason: { type: string }
        ambiguous: { type: boolean }
        first_seen: { type: string, format: date-time }
        last_seen: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        expired: { type: boolean }
        event_ids: { type: array, items: { type: string } }
    RouterScoreComponent:
      type: object
      required: [source_family, evidence_id, score, explanation]
      properties:
        source_family: { type: string }
        evidence_id: { type: string }
        score: { type: integer }
        explanation: { type: string }
    RouterAuthBinding:
      type: object
      required: [session_id, account_id, assigned_ips, source, match_basis, ambiguous]
      properties:
        session_id: { type: string }
        account_id: { type: string }
        assigned_ips: { type: array, items: { type: string } }
        mac: { type: string }
        vlan: { type: string }
        nas_ip: { type: string }
        access_id: { type: string }
        source: { type: string }
        match_basis: { type: string, enum: [exact_mac, exact_endpoint] }
        ambiguous: { type: boolean }
        started_at: { type: string, format: date-time }
        last_confirmed_at: { type: string, format: date-time }
    RouterAssessment:
      type: object
      required: [assessment_id, role, status, confidence, independent_sources, sources, infrastructure, brand_reference_only, association_quality, ambiguous, confirmed_router, rule_version, first_seen, last_seen, expires_at, conflicts, score_components]
      properties:
        assessment_id: { type: string }
        endpoint_id: { type: string }
        ip: { type: string }
        mac: { type: string }
        vlans: { type: array, items: { type: string } }
        brand: { type: string }
        series: { type: string }
        model: { type: string }
        role: { type: string, enum: [router, ap, switch, firewall, endpoint, unknown] }
        status: { type: string, enum: [candidate, likely, confirmed] }
        confidence: { type: integer, minimum: 0, maximum: 100 }
        independent_sources: { type: integer }
        sources: { type: array, items: { type: string } }
        infrastructure: { type: boolean }
        brand_reference_only: { type: boolean }
        association_quality: { type: string }
        ambiguous: { type: boolean }
        confirmed_router: { type: boolean }
        rule_version: { type: string }
        first_seen: { type: string, format: date-time }
        last_seen: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        conflicts: { type: array, items: { type: string } }
        score_components: { type: array, items: { $ref: '#/components/schemas/RouterScoreComponent' } }
        evidence: { type: array, items: { $ref: '#/components/schemas/RouterEvidence' } }
        auth_bindings: { type: array, items: { $ref: '#/components/schemas/RouterAuthBinding' }, description: Current authentication sessions linked only by exact MAC or endpoint identity. These bindings do not affect router confidence. }
        merged_records: { type: integer, minimum: 1, description: Number of active incremental assessments represented by this device row. }
    RouterAssessmentHistory:
      type: object
      required: [status, confidence, rule_version, changed_at, conflicts]
      properties:
        status: { type: string, enum: [candidate, likely, confirmed] }
        confidence: { type: integer, minimum: 0, maximum: 100 }
        rule_version: { type: string }
        changed_at: { type: string, format: date-time }
        conflicts: { type: array, items: { type: string } }
    RouterObservationDetail:
      allOf:
        - { $ref: '#/components/schemas/RouterAssessment' }
        - type: object
          required: [history, current]
          properties:
            current: { type: boolean, description: Whether the stored verdict has unexpired supporting facts; false marks a historical observation. }
            history: { type: array, items: { $ref: '#/components/schemas/RouterAssessmentHistory' } }
    RouterAssessmentPage:
      type: object
      required: [items, page]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/RouterAssessment' } }
        page:
          type: object
          required: [limit, next_cursor, total]
          properties:
            limit: { type: integer }
            next_cursor: { type: [string, 'null'] }
            total: { type: integer }
...